requirements.txt
python-telegram-bot==20.3
pandas
numpy
requests
httpx==0.24.1
